requirements.txt
fastapi==0.110.0
python-multipart
pydantic
groq
python-dotenv
openai
PyPDF2
python-docx
pytesseract
pdf2image
pillow